agregando categoria

+++ sistema/registro_producto.php
@@ -7,20 +7,54 @@ if($_SESSION["rol"]!= 1 && $_SESSION["rol"]!= 2){
 }
 
 
+
 //mouse herramienta misteriosa que nos servira para los filtros 
 //     $date = date_create('2023-04-09 17:01:08');
 //echo date_format($date,'d-m-Y');
 
 include "../conexion.php";
 
+$hashmap = array();
+$opciones = "";
+
+$query_proveedor = mysqli_query($connection,"Select codproveedor,proveedor from proveedor where estatus = 1 order by proveedor asc");
+$result_proveedor = mysqli_num_rows($query_proveedor);
+if($result_proveedor > 0){
+	while($proveedor = mysqli_fetch_array($query_proveedor)){
+		$hashmap[$proveedor["codproveedor"]] = $proveedor["proveedor"];
+		
+			$opciones .=	'<option value="'.$proveedor["codproveedor"].'">'.$proveedor["proveedor"].'</option>';
+		
+	}
+}
+
+$hashmapCategoria = array();
+$opcionesCategoria = "";
+
+$query_categoria = mysqli_query($connection,"Select idcategoria,nombrecat from categoria where estatus = 1 order by nombrecat asc");
+$result_categoria = mysqli_num_rows($query_categoria);
+if($result_categoria > 0){
+	while($categoria = mysqli_fetch_array($query_categoria)){
+		$hashmapCategoria[$categoria["idcategoria"]] = $categoria["nombrecat"];
+		
+			$opcionesCategoria .=	'<option value="'.$categoria["idcategoria"].'">'.$categoria["nombrecat"].'</option>';
+		
+	}
+}
+
+
+
+
+
 	if(!empty($_POST)){
 		$alert = "";
 		if(empty($_POST["proveedor"]) || empty($_POST["descripcion"]) || empty($_POST["precio"]) 
-		 || empty($_POST["cantidad"])  || empty($_POST["medida"]) )
+		 || empty($_POST["cantidad"])  || empty($_POST["medida"]) || empty($_POST["categoria"]) )
 		{
 
-			$alert = "<p class='msg_error'> Los campos Proveedor,Descripcipon,Precio y Cantidad son Obligatorios </p>";
+			$alert = "<p class='msg_error'>" .$_POST["proveedor"]." Los campos Proveedor,Descripcipon,Precio y Cantidad son Obligatorios </p>";
 
+		
 		}else{
 		
 			
@@ -29,6 +63,14 @@ include "../conexion.php";
 				$precio = $_POST["precio"];
 				$cantidad = $_POST["cantidad"];
 				$medida = $_POST["medida"];
+				$adicion = $_POST["adicion"];
+				$idcategoria = $_POST["categoria"];
+
+				if(empty($adicion)){
+					$adicion = 0.00;
+				}
+
+				
 
 				if($medida != 1 && $medida != 2){
 					$medida = 1;
@@ -51,7 +93,11 @@ include "../conexion.php";
 				}
 
 
-						$query_insert = mysqli_query($connection,"INSERT INTO producto(proveedor,descripcion,precio,existencia,codmedida,usuario_id,foto) values($proveedor,'$descripcion','$precio','$cantidad',$medida, $idusuario,'$imgProducto') ");
+					if(!array_key_exists($proveedor,$hashmap) || !array_key_exists($idcategoria,$hashmapCategoria)){
+						$alert = "<p class='msg_error'> No seas chistoso sobrino</p>";
+					}else{
+
+						$query_insert = mysqli_query($connection,"INSERT INTO producto(proveedor,descripcion,precio,existencia,adicion,idcategoria,codmedida,usuario_id,foto) values($proveedor,'$descripcion','$precio','$cantidad',$adicion, $idcategoria, $medida, $idusuario,'$imgProducto') ");
 						if($query_insert){
 									if($nombre_foto != ""){
 										move_uploaded_file($url_temp,$src);
@@ -63,7 +109,8 @@ include "../conexion.php";
 	
 						}else{
 							$alert = "<p class='msg_error'> Error al guardar el Producto</p>";
-						}							
+						}	
+					}						
 				
 		}
 
@@ -96,30 +143,20 @@ include "../conexion.php";
 
 					<form action="" method="post" enctype="multipart/form-data">
                         <label for="proveedor">Proveedor</label>
-
-							<?php
-							
-								$query_proveedor = mysqli_query($connection,"Select codproveedor,proveedor from proveedor where estatus = 1 order by proveedor asc");
-
-									$result_proveedor = mysqli_num_rows($query_proveedor);
-								
-							?>
-
 						<select name="proveedor" id="proveedor">
 							
-						<?php   
-						
-							if($result_proveedor > 0){
-								while($proveedor = mysqli_fetch_array($query_proveedor)){
-									?>
-											<option value="<?php echo $proveedor["codproveedor"]?>"><?php echo $proveedor["proveedor"]?></option>
-									<?php
-								}
-							}
+						<?php echo $opciones?>
 
-						?>
 						</select>
-						
+
+						<label for="categoria">Categoria</label>
+						<select name="categoria" id="categoria">
+							
+						<?php echo $opcionesCategoria?>
+
+						</select>
+
+
 						<label for="producto">Producto</label>
 						<input type="text" name="descripcion" id="descripcion" placeholder="Descripción del Producto">
 
@@ -152,6 +189,9 @@ include "../conexion.php";
 
 						?>
 						</select>
+
+						<label for="adicion">Precio Adicional</label>
+						<input type="number" name="adicion" min="0.00" step="0.01" id="adicion" placeholder="Precio adicional del Producto">
 
 						<div class="photo">
 							<label for="foto">Foto</label>
